#35 creating login tracking state
