+++ src/components/SignOutButton/index.js
@@ -1,0 +1,2 @@
+import { SignOutButton } from "./SignOutButton";
+export default SignOutButton;
