+++ src/components/index.js
@@ -11,6 +11,7 @@ import Header from "./Header";
 import ImageGallery from "./ImageGallery";
 import Navigation from "./Navigation";
 import RecipeListCard from "./RecipeListCard";
+import SignOutButton from "./SignOutButton";
 import SubPageHeader from "./SubPageHeader";
 import ToastNotification from "./ToastNotification";
 export {
@@ -25,6 +26,7 @@ export {
 	ImageGallery,
 	Navigation,
 	RecipeListCard,
+	SignOutButton,
 	SubPageHeader,
 	ToastNotification,
 };
